requirements.txt
Pillow
NumPy
uwsgi
markdown
psycopg2
pygments
hashids
requests
lxml
six
django==1.9.9
django-finalware
html5lib==0.9999999
django-cms
djangocms-cascade
djangorestframework
django-extensions
django-filter
django-classy-tags
django-treebeard
django-sekizai
djangocms-admin-style
django-filer
cmsplugin-filer
django-reversion
djangocms-text-ckeditor
djangocms-link
djangocms-snippet
djangocms-style
djangocms-column
djangocms-grid
djangocms-oembed
djangocms-table
djangocms-blog

